Cleanup and file name refactor. Add Test projects.
