+++ .gitignore
@@ -8,6 +8,21 @@ Desktop.ini
 # Recycle Bin used on file shares
 $RECYCLE.BIN/
 
+# User-specific files
+*.suo
+*.user
+*.sln.docstates
+app.config
+
+# Build results
+
+[Dd]ebug/
+[Rr]elease/
+x64/
+build/
+[Bb]in/
+[Oo]bj/
+
 # Windows Installer files
 *.cab
 *.msi
