+++ App.config
@@ -4,14 +4,6 @@
     <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net" />
   </configSections>
   <log4net>
-    <!--<appender name="LeAppender" type="log4net.Appender.LogentriesAppender, LogentriesLog4net">
-      <Debug value="true" />
-      <HttpPut value="false" />
-      <Ssl value="false" />
-      <layout type="log4net.Layout.PatternLayout">
-        <param name="ConversionPattern" value="%d{ddd MMM dd HH:mm:ss zzz yyyy} %logger %: %level%, %m, " />
-      </layout>
-    </appender>-->
     <appender name="LeAppender" type="log4net.Appender.LogentriesAppender, LogentriesLog4net">
       <immediateFlush value="true" />
       <debug value="true" />
