+++ Google/GoogleImageGetter.cs
@@ -6,7 +6,7 @@ using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TwilioFetchBot.ImageGetter
+namespace TwilioFetchBot.Google
 {
     public class GoogleImageGetter
     {
