+++ TwilioFetchBot.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 14.0.25123.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwilioFetchBot", "TwilioFetchBot.csproj", "{C80A06AA-609A-42CD-BB23-2F7E5EF9643F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwilioFetchBot.Test", "..\TwilioFetchBot.Test\TwilioFetchBot.Test.csproj", "{3FF9645E-00A2-43F9-BEA2-9E3170268460}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{C80A06AA-609A-42CD-BB23-2F7E5EF9643F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C80A06AA-609A-42CD-BB23-2F7E5EF9643F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C80A06AA-609A-42CD-BB23-2F7E5EF9643F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3FF9645E-00A2-43F9-BEA2-9E3170268460}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3FF9645E-00A2-43F9-BEA2-9E3170268460}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3FF9645E-00A2-43F9-BEA2-9E3170268460}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3FF9645E-00A2-43F9-BEA2-9E3170268460}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
