+++ WebAPI Models/Responses/ImgurResponseModels/GalleryImages.cs
