+++ WebAPI Models/Tokens/ImgurCredentials.cs
